Refactoring: move common methods clear_folder() and makeFileGlob() out of each module and into DataIO
