+++ src/GUSTO_Pipeline/DataIO.py
@@ -1,15 +1,44 @@
 """
-GUSTO Pipeline data file handling class for GUSTO L1 SDFITS
+GUSTO Pipeline data file handling class for GUSTO SDFITS
 """
 
 import warnings
 from astropy.io import fits
 from astropy.utils.exceptions import AstropyWarning
+import os
+import shutil
+import glob
 import numpy as np
 import numpy.ma as ma
 
 warnings.filterwarnings('ignore', category=Warning,
                         message=' FITSFixedWarning: ', append=True)
+
+
+def clear_folder(folder_path):
+    print('Erasing contents of '+folder_path)
+    for filename in os.listdir(folder_path):
+        file_path = os.path.join(folder_path, filename)
+        try:
+            if os.path.isfile(file_path) or os.path.islink(file_path):
+                os.unlink(file_path)
+            elif os.path.isdir(file_path):
+                shutil.rmtree(file_path)
+        except Exception as e:
+            print(f"Failed to delete {file_path}. Reason: {e}")
+
+            
+def makeFileGlob(inDir, prefix, suffix, scanRange):
+    ignore = [10086, 13638, 17751, 27083, 28089, 4564, 7165, 7167]
+    filter=prefix+'*.'+suffix
+    sdirs = sorted(glob.glob(os.path.join(inDir,filter)))
+    dsc = [int(os.path.split(sdir)[1].split('_')[2].split('.')[0]) for sdir in sdirs]
+    dfiles = []
+    for i,ds in enumerate(dsc):
+        if (ds >= scanRange[0]) & (ds <= scanRange[1]) & (ds not in ignore):
+            dfiles.append(sdirs[i])            
+    return dfiles
+
 
 
 def loadSDFITS(ifile, verbose=False, usemask=False):
