+++ src/GUSTO_Pipeline/L07_telemetry.py
@@ -1,18 +1,16 @@
-#!/usr/bin/env python3.11
 from influxdb import InfluxDBClient
 from astropy.io import fits
 from multiprocessing import Pool
 from functools import partial
-import glob
-import struct
+from struct import unpack
 from datetime import datetime
-import sys
-import os.path
+
+import os
 import numpy as np
-import configargparse
-import argparse
 import subprocess
+
 from .flagdefs import *
+from .DataIO import *
 
 
 def runGitLog():
@@ -78,18 +76,6 @@ def makeSequences(input, output):
         print(f"{seqNum:05d}", f"{startID:05d}", f"{endID:05d}", seqType, obj, file=seqFile)
 
 
-def makeFileGlob(inDir, prefix, suffix, scanRange):
-    ignore = [10086, 13638, 17751, 27083, 28089, 4564, 7165, 7167]
-    filter=prefix+'*.'+suffix
-    sdirs = sorted(glob.glob(os.path.join(inDir,filter)))
-    dsc = [int(os.path.split(sdir)[1].split('_')[2].split('.')[0]) for sdir in sdirs]
-    dfiles = []
-    for i,ds in enumerate(dsc):
-        if (ds >= scanRange[0]) & (ds <= scanRange[1]) & (ds not in ignore):
-            dfiles.append(sdirs[i])            
-    return dfiles
-
-
 def checkSequence(fileList):
     seqFlag = 0
     REF=0
@@ -149,7 +135,7 @@ def makeUDP(startID, stopID, dir):
                     chunk = stream.read(BUFSIZE)
                     if not chunk:
                         break 
-                    data=struct.unpack('<2I5f2I15d',chunk)
+                    data=unpack('<2I5f2I15d',chunk)
                     output.append(list(data))  # convert tuple to list to modify the time
         except Exception as error:
             print(error)
@@ -409,19 +395,6 @@ def processFITS(data_path, input_files, output_file, bandNum, pointingStream, se
     hduList = fits.HDUList([fits.PrimaryHDU(header=header), hdu])
     hduList.writeto(output_file, overwrite=True)
     return
-
-
-def clear_folder(folder_path):
-    print('Erasing contents of '+folder_path)
-    for filename in os.listdir(folder_path):
-        file_path = os.path.join(folder_path, filename)
-        try:
-            if os.path.isfile(file_path) or os.path.islink(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                shutil.rmtree(file_path)
-        except Exception as e:
-            print(f"Failed to delete {file_path}. Reason: {e}")
 
 
 def scanSequenceFile(input, options):
