+++ src/GUSTO_Pipeline/L08_chanflags.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python3.11
-
 import os
 import sys
 import glob
@@ -13,23 +11,7 @@ from multiprocessing import Pool
 from functools import partial
 import subprocess
 from .flagdefs import *
-
-
-def clear_folder(folder_path):
-    print('Erasing contents of '+folder_path)
-    for filename in os.listdir(folder_path):
-        file_path = os.path.join(folder_path, filename)
-        try:
-            if os.path.isfile(file_path) or os.path.islink(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                shutil.rmtree(file_path)
-        except Exception as e:
-            print(f"Failed to delete {file_path}. Reason: {e}")
-
-            
-def flatten(xss):
-    return [x for xs in xss for x in xs]
+from .DataIO import *
 
 
 def runGitLog():
@@ -51,18 +33,6 @@ def find_mixer(line, mixer):
         return mixers.index(mixer)
     else:
         return 5
-
-    
-def makeFileGlob(inDir, prefix, suffix, scanRange):
-    ignore = [10086, 13638, 17751, 27083, 28089, 4564, 7165, 7167]
-    filter=prefix+'*.'+suffix
-    sdirs = sorted(glob.glob(os.path.join(inDir,filter)))
-    dsc = [int(os.path.split(sdir)[1].split('_')[2].split('.')[0]) for sdir in sdirs]
-    dfiles = []
-    for i,ds in enumerate(dsc):
-        if (ds >= scanRange[0]) & (ds <= scanRange[1]) & (ds not in ignore):
-            dfiles.append(sdirs[i])            
-    return dfiles
 
 
 def L08_Pipeline(args, scanRange):
