+++ src/GUSTO_Pipeline/L09_calibrate.py
@@ -2,11 +2,9 @@
 """
 This is the GUSTO L09 Pipeline.
 """
-import glob
 import numpy as np
 import numpy.ma as ma
-import time
-import sys
+#import time
 import os
 import subprocess
 import logging
@@ -17,7 +15,7 @@ from astropy.io import fits
 from multiprocessing.pool import Pool
 from PyAstronomy import pyasl
 
-from .DataIO import loadSDFITS
+from .DataIO import *
 from .Logger import *
 from .Configuration import *
 from .flagdefs import *
@@ -31,19 +29,6 @@ def runGitLog():
         return f"Error: {e}"
 
 
-def clear_folder(folder_path):
-    print('Erasing contents of '+folder_path)
-    for filename in os.listdir(folder_path):
-        file_path = os.path.join(folder_path, filename)
-        try:
-            if os.path.isfile(file_path) or os.path.islink(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                shutil.rmtree(file_path)
-        except Exception as e:
-            print(f"Failed to delete {file_path}. Reason: {e}")
-
-    
 def despike_polyRes(x, data, cflags, start, stop, points=60, count=3, deg=2, dx=1, stdlim=4.0):
     mask = np.zeros(len(data), dtype=bool)
     iin, iout = pyasl.slidingPolyResOutlier(x[start:stop], data[start:stop], points=points, count=count, deg=deg, stdlim=stdlim, controlPlot=False, dx=dx, mode='both')
@@ -377,21 +362,7 @@ def checkRowflag(rowflagvalue, rowflagfilter=0):
     negfilt = ~rowflagfilter#.__invert__()
     return np.array([False if int(negfilt).__and__(int(rfvalue)) else True for rfvalue in rowflagvalue], dtype=bool)
 
-
 checkRowflags = np.vectorize(checkRowflag)
-
-
-def makeFileGlob(inDir, prefix, suffix, scanRange):
-    ignore = [10086, 13638, 17751, 27083, 28089, 4564, 7165, 7167]
-    filter=prefix+'*.'+suffix
-    sdirs = sorted(glob.glob(os.path.join(inDir,filter)))
-    dsc = [int(os.path.split(sdir)[1].split('_')[2].split('.')[0]) for sdir in sdirs]
-    dfiles = []
-    for i,ds in enumerate(dsc):
-        if (ds >= scanRange[0]) & (ds <= scanRange[1]) & (ds not in ignore):
-            dfiles.append(sdirs[i])            
-    return dfiles
-
 
 
     #########################################################
