+++ src/GUSTO_Pipeline/L10_pointing.py
@@ -3,10 +3,8 @@
 This is the source file for the GUSTO L1.0 Pipeline step which corrects the per-pixel
 pointing and converts the frequency axis to Doppler velocity in the LSR frame of reference.
 """
-import glob
 import numpy as np
 import time
-import sys
 import os
 import subprocess
 from datetime import datetime
@@ -18,7 +16,7 @@ from astropy import units as u
 from astropy.coordinates import SkyCoord, EarthLocation, AltAz
 from astropy.time import Time
 
-from .DataIO import loadSDFITS
+from .DataIO import *
 from .Logger import *
 
 offsetfile0 = files('GUSTO_Pipeline') / 'calib/offsets.txt'
@@ -29,31 +27,6 @@ def runGitLog():
         return result.stdout
     except subprocess.CalledProcessError as e:
         return f"Error: {e}"
-
-
-def clear_folder(folder_path):
-    print('Erasing contents of '+folder_path)
-    for filename in os.listdir(folder_path):
-        file_path = os.path.join(folder_path, filename)
-        try:
-            if os.path.isfile(file_path) or os.path.islink(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                shutil.rmtree(file_path)
-        except Exception as e:
-            print(f"Failed to delete {file_path}. Reason: {e}")
-
-            
-def makeFileGlob(inDir, prefix, suffix, scanRange):
-    ignore = [10086, 13638, 17751, 27083, 28089, 4564, 7165, 7167]
-    filter=prefix+'*.'+suffix
-    sdirs = sorted(glob.glob(os.path.join(inDir,filter)))
-    dsc = [int(os.path.split(sdir)[1].split('_')[2].split('.')[0]) for sdir in sdirs]
-    dfiles = []
-    for i,ds in enumerate(dsc):
-        if (ds >= scanRange[0]) & (ds <= scanRange[1]) & (ds not in ignore):
-            dfiles.append(sdirs[i])            
-    return dfiles
 
 
 def L10_Pipeline(args, scanRange, verbose=False):
@@ -127,10 +100,6 @@ def processL10(params, verbose=True):
     global commit_info
 
     line, inDir, outDir, dfile, debug = params[0], params[1], params[2], params[3], params[4]
-    
-    # define some processing data first (maybe relocat to function later?)
-
-    #logger.info('loading: ', os.path.join(inDir,dfile), ' for line: ', line)
     spec, data, hdr, hdr1 = loadSDFITS(os.path.join(inDir,dfile), verbose=False)
     
     umixers = np.unique(data['MIXER'])
